[EC-298] Moved benchmark tests from test to  benchmark task
