+++ build.sbt
@@ -45,15 +45,18 @@ val dep = {
 
 val Integration = config("it") extend Test
 
+val Benchmark = config("benchmark") extend Test
+
 val Evm = config("evm") extend Test
 
 val Ets = config("ets") extend Test
 
 val root = project.in(file("."))
-    .configs(Integration, Evm, Ets)
+    .configs(Integration, Benchmark, Evm, Ets)
     .settings(commonSettings: _*)
     .settings(libraryDependencies ++= dep)
     .settings(inConfig(Integration)(Defaults.testSettings) : _*)
+    .settings(inConfig(Benchmark)(Defaults.testSettings) : _*)
     .settings(inConfig(Evm)(Defaults.testSettings) : _*)
     .settings(inConfig(Ets)(Defaults.testSettings) : _*)
